refactoring: lce/python -> lce/tf/python
