+++ larq_compute_engine/__init__.py
@@ -1,10 +1,10 @@
-from larq_compute_engine.python.ops.compute_engine_ops import bsign
-from larq_compute_engine.python.ops.compute_engine_ops import (
+from larq_compute_engine.tf.python.ops.compute_engine_ops import bsign
+from larq_compute_engine.tf.python.ops.compute_engine_ops import (
     bconv2d,
     bconv2d8,
     bconv2d32,
     bconv2d64,
 )
-from larq_compute_engine.python.modelconverter import ModelConverter
+from larq_compute_engine.tf.python.modelconverter import ModelConverter
 
 __all__ = ["bsign", "bconv2d", "bconv2d8", "bconv2d32", "bconv2d64", "ModelConverter"]
