+++ larq_compute_engine/tf/python/ops/compute_engine_ops.py
