+++ larq_compute_engine/tf/python/utils.py
